requirements.txt
# Core dependencies
pydantic>=2.6,<3.0
pytest>=8.0,<9.0
httpx>=0.24.0,<1.0
beautifulsoup4>=4.10.0,<5.0
ddgs>=9.0
pypdf>=3.0.0,<4.0
trafilatura>=1.6.0,<2.0

# Web framework
fastapi>=0.104.0,<1.0
uvicorn[standard]>=0.24.0,<1.0
websockets>=12.0,<13.0

# Development tools
pytest-cov>=4.0.0,<5.0
black>=23.0.0,<24.0
isort>=5.12.0,<6.0
mypy>=1.5.0,<2.0

# Optional dependencies
aiohttp>=3.8.0,<4.0
python-multipart>=0.0.6,<1.0

# Authentication
bcrypt>=4.0.0,<5.0
PyJWT>=2.8.0,<3.0
cryptography>=41.0.0,<42.0

# Document Q&A System dependencies
faiss-cpu>=1.7.0,<2.0
sentence-transformers>=2.2.0,<3.0
python-docx>=0.8.11,<1.0
PyMuPDF>=1.22.0,<2.0
